python/requirements-desktop.txt
# HermesDesk pruned Python dependencies.
#
# Mirrors the [project] core deps from upstream pyproject.toml, minus
# the heavy/incompatible extras (RL, voice STT, multi-platform messaging,
# matrix, etc.). Keep version pins aligned with upstream.

# --- Core (mirrors upstream [project].dependencies) -----------------------
openai>=2.21.0,<3
anthropic>=0.39.0,<1
python-dotenv>=1.2.1,<2
fire>=0.7.1,<1
httpx[socks]>=0.28.1,<1
rich>=14.3.3,<15
tenacity>=9.1.4,<10
pyyaml>=6.0.2,<7
requests>=2.33.0,<3
jinja2>=3.1.5,<4
pydantic>=2.12.5,<3
prompt_toolkit>=3.0.52,<4

# --- Tools we keep --------------------------------------------------------
exa-py>=2.9.0,<3                 # web search
firecrawl-py>=4.16.0,<5          # web scrape
parallel-web>=0.4.2,<1
fal-client>=0.13.1,<1            # image gen
edge-tts>=7.2.7,<8               # free TTS

# --- Skills hub bot identity (PyJWT for GitHub App auth) -----------------
PyJWT[crypto]>=2.12.0,<3

# --- Web server (Hermes' built-in localhost UI) --------------------------
fastapi>=0.104.0,<1
uvicorn[standard]>=0.24.0,<1

# --- HermesDesk L1 built-in helpers (QuickActions) ------------------------
openpyxl>=3.1.2,<4
python-docx>=1.1.2,<2
pypdf>=5.1.0,<6
Pillow>=11.0.0,<12

# --- Hidden behind Power-user toggle (still installed; importable when toggled) ---
mcp>=1.2.0,<2

# --- Explicitly NOT installed in v1 ---------------------------------------
# faster-whisper      ~200 MB; cloud STT only at v1
# sounddevice         pulls portaudio; deferred with voice mode
# numpy>=1.24         only needed by voice; deferred
# python-telegram-bot ; gateway/ stripped
# discord.py          ; gateway/ stripped
# slack-bolt          ; gateway/ stripped
# mautrix             ; gateway/ stripped
# modal               ; remote terminal backend stripped
# daytona             ; remote terminal backend stripped
# atroposlib, tinker  ; RL stripped
# yc-bench            ; benchmark stripped
# debugpy, pytest     ; dev-only, not bundled
